+++ NEWS.md
@@ -53,6 +53,7 @@
 1. Coercion to/from integer64 is expanded greatly (includes #199). Thanks @hcirellu.
    - `as.Date`, `as.POSIXct`, `as.POSXlt`, `as.complex`, and `as.raw` get an `integer64` method.
    - `as.integer64` gets `Date`, `POSIXct`, `POSXlt`, `complex`, `raw`, and `difftime` methods.
+1. `sortcache`, `sortordercache` and `ordercache` get a new argument `na.last`.
 
 ## BUG FIXES